task_manager/requirements-web.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
gunicorn
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
pydantic[email]
pydantic-settings
sqlalchemy
databases
alembic
python-multipart
redis
celery[redis]
psycopg2-binary
authlib
itsdangerous
boto3
beautifulsoup4
cryptography
python-dotenv
requests
pgvector
google-generativeai
numpy
google-cloud-vision
nltk
rapidfuzz
pdfplumber
langchain-chroma>=0.1.2
chromadb>=0.5
PyMuPDF
langchain==0.3.27
langchain-huggingface
sentence-transformers
